Kotlin based Data Loader Yay!

+++ ingest-service/server.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"log"
 	"os"
@@ -9,7 +10,10 @@ import (
 	"strings"
 	"time"
 
+	"database/sql"
+
 	"github.com/gin-gonic/gin"
+	_ "github.com/lib/pq"
 	"github.com/nats-io/nats.go"
 	"github.com/saiprasadkrishnamurthy/ingest-service/handler"
 	"github.com/saiprasadkrishnamurthy/ingest-service/listener"
@@ -33,48 +37,74 @@ func init() {
 
 func main() {
 	r := gin.Default()
-	serviceFactory := setupDependencies()
+	readConfig()
+
+	// Postgres DB
+	host := viper.GetString("db_host")
+	port := viper.GetInt32("db_port")
+	user := viper.GetString("db_user")
+	password := viper.GetString("db_password")
+	dbname := viper.GetString("db_dbname")
+	psqlconn := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable", host, port, user, password, dbname)
+	db, err := sql.Open("postgres", psqlconn)
+	CheckError(err)
+
+	// close database
+	defer db.Close()
+
+	// check db
+	err = db.Ping()
+	CheckError(err)
+	serviceFactory := setupDependencies(db)
 	apiVersion := viper.GetString("api_version")
 	serviceName := viper.GetString("service_name")
 	r.Use(middleware.Auth(serviceFactory.MongoClient))
 	r.Use(middleware.CORSMiddleware())
 	r.POST(serviceName+"/api/"+apiVersion+"/ingest/:name/", func(ctx *gin.Context) {
+		form, _ := ctx.MultipartForm()
 		name := ctx.Param("name")
-		idField := ctx.Query("idField")
 		tags := ctx.Query("tags")
 		file, _ := ctx.FormFile("file")
-		destDir := uuid.NewV4().String()
-		defer os.RemoveAll(destDir)
-		defer os.Remove(destDir)
+		dataDefinitions := strings.ReplaceAll(form.Value["dataDefinitions"][0], "\\r\\n", "")
 
-		if err := os.Mkdir(destDir, os.ModePerm); err != nil {
-			log.Fatal(err)
+		var def model.DataDefinitions
+		e := json.Unmarshal([]byte(dataDefinitions), &def)
+
+		if e != nil {
+			ctx.JSON(400, model.PartitionFileResponse{Errors: []model.ApiError{{Code: "ING0001", Message: "Invalid data definition"}}})
+		} else {
+			destDir := uuid.NewV4().String()
+			defer os.RemoveAll(destDir)
+			defer os.Remove(destDir)
+
+			if err := os.Mkdir(destDir, os.ModePerm); err != nil {
+				log.Fatal(err)
+			}
+
+			err = ctx.SaveUploadedFile(file, destDir+"/"+file.Filename)
+
+			if err != nil {
+				fmt.Println("Cant save the uploaded file to disk! ", err)
+			}
+
+			tenant, _ := ctx.Get("tenant")
+			objectStorageId, _ := ctx.Get("objectStorageId")
+			response := serviceFactory.HandlePartition(
+				&model.PartitionFileRequest{
+					Tenant:          tenant.(string),
+					ObjectStorageId: objectStorageId.(string),
+					Name:            name,
+					Tags:            strings.Split(tags, ","),
+					ZipFilePath:     destDir + "/" + file.Filename,
+					DataDefinitions: def,
+				})
+			ctx.JSON(200, response)
 		}
-
-		err := ctx.SaveUploadedFile(file, destDir+"/"+file.Filename)
-
-		if err != nil {
-			fmt.Println("Cant save the uploaded file to disk! ", err)
-		}
-
-		tenant, _ := ctx.Get("tenant")
-		objectStorageId, _ := ctx.Get("objectStorageId")
-		response := serviceFactory.HandlePartition(
-			&model.PartitionFileRequest{
-				IdField:         idField,
-				Tenant:          tenant.(string),
-				ObjectStorageId: objectStorageId.(string),
-				Name:            name,
-				Tags:            strings.Split(tags, ","),
-				ZipFilePath:     destDir + "/" + file.Filename,
-			})
-		ctx.JSON(200, response)
 	})
 	r.Run(":" + strconv.Itoa(viper.GetInt("server_port")))
 }
 
-func setupDependencies() *handler.ServiceFactory {
-	readConfig()
+func setupDependencies(db *sql.DB) *handler.ServiceFactory {
 	log := &model.Log{
 		Info:  log.New(logFile, "INFO: ", log.Ldate|log.Ltime|log.Lshortfile),
 		Warn:  log.New(logFile, "WARNING: ", log.Ldate|log.Ltime|log.Lshortfile),
@@ -83,7 +113,9 @@ func setupDependencies() *handler.ServiceFactory {
 	fmt.Println(" Log File ===> ", logFile.Name())
 	return NewServiceFactory(log,
 		service.NewIngestService(log),
-		storage.NewS3Manager(log))
+		storage.NewS3Manager(log),
+		db,
+	)
 }
 
 func readConfig() {
@@ -107,7 +139,8 @@ func readConfig() {
 }
 func NewServiceFactory(log *model.Log,
 	ingestService *service.IngestService,
-	s3Manager *storage.S3Manager) *handler.ServiceFactory {
+	s3Manager *storage.S3Manager,
+	db *sql.DB) *handler.ServiceFactory {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 	client, err := mongo.Connect(ctx, options.Client().ApplyURI(viper.GetString("mongo_uri")))
@@ -132,12 +165,19 @@ func NewServiceFactory(log *model.Log,
 			viper.GetString("ingest_event_subject"), x.Created)
 	}
 
-	listener.InitializeAllListeners(nc, log, client, s3Manager)
+	listener.InitializeAllListeners(nc, log, client, s3Manager, db)
 	return &handler.ServiceFactory{
 		IngestService: ingestService,
 		MongoClient:   client,
 		Log:           log,
 		S3Manager:     s3Manager,
 		Nats:          nc,
+		DB:            db,
+	}
+
+}
+func CheckError(err error) {
+	if err != nil {
+		panic(err)
 	}
 }
